# docs/superpowers/plans/2026-06-17-api-spec.yaml
openapi: "3.0.3"
info:
  title: Intelligent Foreign Trade Financial API
  description: |
    API for the China foreign trade financial software (智能外贸财务系统).
    Features: AI-powered voucher generation, multi-ledger accounting, period closing,
    OEM commission calculation, print & export, audit trail.

    **Auth:** JWT Bearer token in `Authorization` header. CSRF via `X-CSRF-Token`.
    **Multi-ledger:** Pass `X-Ledger-Id` header to scope all operations to a ledger.
  version: "1.0.0"
  contact:
    name: Trad Account Team
servers:
  - url: http://localhost:8004
    description: Local development (8000-8003 blocked by zombie ports on Windows)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string, format: password }
    LoginResponse:
      type: object
      properties:
        access_token: { type: string }
        token_type: { type: string, example: bearer }
        user: { $ref: '#/components/schemas/User' }
...
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        role: { type: string }
        is_active: { type: boolean }
        last_login: { type: string, format: date-time }
        created_at: { type: string, format: date-time }

    # --- Ledger ---
    LedgerCreate:
      type: object
      required: [name, start_year, start_month]
      properties:
        name: { type: string, description: 账套名称 }
        company_name: { type: string, description: 公司全称 }
        base_currency: { type: string, default: CNY }
        start_year: { type: integer }
        start_month: { type: integer }
        llm_provider: { type: string, default: deepseek }
        llm_api_key: { type: string }
        llm_base_url: { type: string }
        llm_model_name: { type: string, default: deepseek-chat }
    Ledger:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        company_name: { type: string }
        base_currency: { type: string }
        start_year: { type: integer }
        start_month: { type: integer }
        is_active: { type: boolean }
        llm_provider: { type: string }
        llm_model_name: { type: string }
        created_at: { type: string, format: date-time }

    # --- Account ---
    AccountCreate:
      type: object
      required: [code, name, account_type, balance_direction]
      properties:
        code: { type: string, description: 科目编码, example: "1001" }
        name: { type: string, description: 科目名称, example: "库存现金" }
        parent_id: { type: integer, nullable: true }
        account_type: { type: string, enum: [ASSET, LIABILITY, EQUITY, COST, PROFIT_LOSS] }
        balance_direction: { type: string, enum: [DEBIT, CREDIT] }
        opening_balance: { type: number, default: 0.00 }
        is_active: { type: boolean, default: true }
    Account:
      type: object
      properties:
        id: { type: integer }
        ledger_id: { type: integer }
        code: { type: string }
        name: { type: string }
        parent_id: { type: integer, nullable: true }
        account_type: { type: string }
        balance_direction: { type: string }
        opening_balance: { type: number }
        is_active: { type: boolean }
        children: { type: array, items: { $ref: '#/components/schemas/Account' } }

    # --- Partner ---
    PartnerCreate:
      type: object
      required: [code, name, partner_type]
      properties:
        code: { type: string, description: 往来单位编码 }
        name: { type: string, description: 往来单位名称 }
        partner_type: { type: string, enum: [CUSTOMER, VENDOR, BOTH] }
        is_active: { type: boolean, default: true }
    Partner:
      type: object
      properties:
        id: { type: integer }
        ledger_id: { type: integer }
        code: { type: string }
        name: { type: string }
        partner_type: { type: string }
        is_active: { type: boolean }

    # --- Voucher ---
    VoucherEntrySchema:
      type: object
      required: [account_code, summary, direction, amount]
      properties:
        account_code: { type: string }
        summary: { type: string, description: 摘要 }
        direction: { type: string, enum: ["借", "贷"] }
        amount: { type: number, exclusiveMinimum: 0 }
        currency_code: { type: string, default: CNY }
        original_amount: { type: number, nullable: true }
        exchange_rate: { type: number, default: 1.0000 }
        partner_id: { type: integer, nullable: true }
    VoucherCreate:
      type: object
      required: [voucher_date, entries]
      properties:
        voucher_date: { type: string, format: date }
        voucher_number: { type: string, description: 凭证字号, auto-generated if omitted }
        attachments_count: { type: integer, default: 0 }
        contract_number: { type: string, nullable: true, description: OEM合同号 }
        entries: { type: array, items: { $ref: '#/components/schemas/VoucherEntrySchema' }, minItems: 2 }
    VoucherEntryResponse:
      type: object
      properties:
        id: { type: integer }
        account_id: { type: integer }
        account: { type: object, properties: { id: { type: integer }, code: { type: string }, name: { type: string } } }
        summary: { type: string }
        direction: { type: string }
        amount: { type: number }
        currency_code: { type: string }
        original_amount: { type: number, nullable: true }
        exchange_rate: { type: number }
        partner_id: { type: integer, nullable: true }
        vat_rate: { type: number, nullable: true }
        vat_amount: { type: number, nullable: true }
    VoucherResponse:
      type: object
      properties:
        id: { type: integer }
        ledger_id: { type: integer }
        voucher_number: { type: string }
        voucher_date: { type: string, format: date }
        attachments_count: { type: integer }
        status: { type: string, enum: [DRAFT, APPROVED, POSTED] }
        source_type: { type: string, nullable: true }
        contract_number: { type: string, nullable: true }
        entries: { type: array, items: { $ref: '#/components/schemas/VoucherEntryResponse' } }
...
    ErrorResponse:
      type: object
      properties:
        detail: { type: string }

    # --- Health ---
...
security:
  - BearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [Auth]
      summary: Login
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/LoginResponse' } } } }
        '401': { description: Invalid credentials, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorResponse' } } } }
        '429': { description: Rate limited (5 req/min on login) }
...
  /ledgers:
    get:
      tags: [Ledgers]
      summary: List all ledgers accessible to current user
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Ledger' } } } } }
    post:
      tags: [Ledgers]
      summary: Create a new ledger (账套)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LedgerCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Ledger' } } } }

  # ========== ACCOUNTS ==========
  /accounts:
    get:
      tags: [Accounts]
      summary: List chart of accounts (hierarchical)
      parameters:
        - name: active_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Account' } } } } }
    post:
      tags: [Accounts]
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }
  /accounts/{id}:
    put:
      tags: [Accounts]
      summary: Update an account
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AccountCreate' }
      responses:
        '200': { description: Updated, content: { application/json: { schema: { $ref: '#/components/schemas/Account' } } } }

  # ========== PARTNERS ==========
  /partners:
    get:
      tags: [Partners]
      summary: List business partners
      parameters:
        - name: type
          in: query
          schema: { type: string, enum: [CUSTOMER, VENDOR, BOTH] }
        - name: active_only
          in: query
          schema: { type: boolean, default: true }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Partner' } } } } }
    post:
      tags: [Partners]
      summary: Create a business partner
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PartnerCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Partner' } } } }

  # ========== VOUCHERS ==========
  /vouchers:
    get:
      tags: [Vouchers]
      summary: List vouchers with pagination & filters
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20, maximum: 100 }
        - name: status
          in: query
          schema: { type: string, enum: [DRAFT, APPROVED, POSTED] }
        - name: date_from
          in: query
          schema: { type: string, format: date }
        - name: date_to
          in: query
          schema: { type: string, format: date }
        - name: keyword
          in: query
          schema: { type: string, description: 摘要/凭证号搜索 }
      responses:
        '200': { description: OK, content: { application/json: { schema: { type: object, properties: { items: { type: array, items: { $ref: '#/components/schemas/VoucherResponse' } }, total: { type: integer } } } } } }
    post:
      tags: [Vouchers]
      summary: Create a voucher manually
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VoucherCreate' }
      responses:
        '200': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/VoucherResponse' } } } }
...
  /vouchers/{id}:
    get:
      tags: [Vouchers]
      summary: Get voucher by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/VoucherResponse' } } } }
...